Make the tail follow you.
Instead of leaving the trail behind you forever,
start with a nominal length (~0x20 pixels) and
trim as we go.  When you eat food, the tail can
grow longer.

Also make several different types of food, for
fun.  And adjust the timing for playability.
Don't hit the H key when playing, that's the
"hyper" key and will run you into a wall very fast.